WordAnalyzer deleted
