--- TagCloud/Infrastructure/Text/ITokenAnalyzer.cs
@@ -1,10 +1,0 @@
-using System.Collections.Generic;
-using TagCloud.Infrastructure.Text.Information;
-
-namespace TagCloud.Infrastructure.Text
-{
-    public interface ITokenAnalyzer<TToken>
-    {
-        public IEnumerable<(TToken, TokenInfo)> Analyze(IEnumerable<TToken> words);
-    }
-}
