+++ TagCloud/Program.cs
@@ -24,7 +24,7 @@ namespace TagCloud
         {
             var builder = new ContainerBuilder();
             builder.RegisterType<TxtReader>().As<IReader<string>>();
-            builder.RegisterType<WordAnalyzer>().As<ITokenAnalyzer<string>>();
+            builder.RegisterType<TagCloudGenerator>().As<IImageGenerator>();
 
             builder.RegisterType<LowerCaseConveyor>().As<IConveyor<string>>();
             var myStemPath = GetReleasePath("mystem");
