+++ TagCloudTests/ImageSizeSettingsManagerTests.cs
@@ -17,7 +17,6 @@ namespace TagCloudTests
         {
             var builder = new ContainerBuilder();
             builder.RegisterType<TxtReader>().As<IReader<string>>();
-            builder.RegisterType<WordAnalyzer>().As<ITokenAnalyzer<string>>();
 
             builder.RegisterType<Settings>()
                 .AsSelf()
